Button "Select Users" and check button "Select All Users" Updated
Now the button Select Users open the window to select the users of the database and the button "Select All Users" are working in parts.
The button show the message informing what the all users are selected, but I don't configure it yet to catch all the users and put in the file.